# Apply perf improvements from pydata talk