#### What does this implement or fix?
Speeds up binary operations with 2 sparse columns

New benchmark results:
```
Before change
---------------------------------------------------------------------------------------
Benchmark                                             Time             CPU   Iterations
---------------------------------------------------------------------------------------
BM_single_column_projection/100                     736 ms          736 ms            1
BM_single_column_projection/99                      783 ms          783 ms            1
BM_single_column_projection/90                      694 ms          694 ms            1
BM_single_column_projection/50                      406 ms          406 ms            2
BM_single_column_projection/10                      103 ms          103 ms            7
BM_single_column_projection/1                      13.2 ms         13.2 ms           39
BM_two_column_projection/100                        766 ms          766 ms            1
BM_two_column_projection/99                       40850 ms        40849 ms            1
BM_two_column_projection/90                       33778 ms        33777 ms            1
BM_two_column_projection/50                       10643 ms        10643 ms            1
BM_two_column_projection/10                         530 ms          530 ms            1
BM_two_column_projection/1                         41.0 ms         41.0 ms           16
After change
---------------------------------------------------------------------------------------
Benchmark                                             Time             CPU   Iterations
---------------------------------------------------------------------------------------
BM_single_column_projection/100                     591 ms          591 ms            1
BM_single_column_projection/99                      550 ms          550 ms            1
BM_single_column_projection/90                      648 ms          648 ms            1
BM_single_column_projection/50                      391 ms          391 ms            2
BM_single_column_projection/10                     92.3 ms         92.3 ms            7
BM_single_column_projection/1                      13.4 ms         13.4 ms           59
BM_two_column_projection/100                        775 ms          775 ms            1
BM_two_column_projection/99                        2273 ms         2273 ms            1
BM_two_column_projection/90                        2475 ms         2475 ms            1
BM_two_column_projection/50                        2349 ms         2349 ms            1
BM_two_column_projection/10                         477 ms          477 ms            2
BM_two_column_projection/1                         79.8 ms         79.8 ms            9
```