@maxim-morozov 2025-06-13
Nice speedup, but I guess the numbers depend on the density of the columns